Lighten prereqs that aren't fulfilled
