+++ src/components/ClassInfo.vue
@@ -215,6 +215,7 @@ import SubjectScroll from '../components/SubjectScroll.vue';
 import ExpansionReqs from '../components/ExpansionReqs.vue';
 import colorMixin from './../mixins/colorMixin.js';
 import schedule from './../mixins/schedule.js';
+import reqFulfillment from './../mixins/reqFulfillment.js';
 
 export default {
   name: 'ClassInfo',
@@ -222,7 +223,7 @@ export default {
     'subject-scroll': SubjectScroll,
     'expansion-reqs': ExpansionReqs
   },
-  mixins: [colorMixin, schedule],
+  mixins: [colorMixin, schedule, reqFulfillment],
   data: function () { return {} },
   computed: {
     addingFromCard () {
@@ -248,6 +249,7 @@ export default {
         : { reqs: [] };
     },
     subjectsWithPrereq: function () {
+      console.log('This other computed property is running');
       const currentID = this.currentSubject.subject_id;
       const currentDept = currentID.substring(0, currentID.indexOf('.'));
       var IDMatcher = new RegExp('(^|[^\\da-zA-Z])' + currentID.replace(/[.*+?^${}()|[\]\\]/g, '\\$&') + '(?![\\da-zA-Z])');
@@ -263,6 +265,16 @@ export default {
           return 0;
         }
       });
+    },
+    firstAppearance: function() {
+      const currentSubjects = this.$store.state.roads[this.$store.state.activeRoad].contents.selectedSubjects;
+      const currentID = this.currentSubject.subject_id;
+
+      const subjectInSemesters = currentSubjects.map((semesterSubjects) => {
+        return semesterSubjects.map((subj) => subj.id).indexOf(currentID) >= 0;
+      });
+
+      return subjectInSemesters.indexOf(true);
     }
   },
   methods: {
@@ -280,6 +292,8 @@ export default {
     parseRequirements: function (requirements) {
       // TODO: a way to make this more ETU?
       // remove spaces after commas and slashes
+      const _this = this;
+
       requirements = requirements.replace(/([,/])\s+/g, '$1');
       function getParenGroup (str) {
         if (str[0] === '(') {
@@ -303,6 +317,7 @@ export default {
           return undefined;
         }
       }
+
       function getNextReq (reqString) {
         if (reqString[0] === '(') {
           return getParenGroup(reqString);
@@ -318,12 +333,15 @@ export default {
           }
         }
       }
+
       function isBaseReq (req) {
         return /[/(),]/g.exec(req) === null;
       }
+
       const getClassInfo = this.classInfo;
+
       function parseReqs (reqString) {
-        const parsedReq = { reqs: [], subject_id: '', connectionType: '', title: '', expansionDesc: '', topLevel: false };
+        const parsedReq = { reqs: [], subject_id: '', connectionType: '', title: '', expansionDesc: '', topLevel: false, fulfilled: false };
         let onereq;
         let connectionType;
         let nextConnectionType;
@@ -334,19 +352,31 @@ export default {
           }
           if (isBaseReq(onereq)) {
             if (onereq.indexOf("'") >= 0) {
-              parsedReq.reqs.push({ subject_id: onereq.replace(/'/g, ''), title: '' });
+              parsedReq.reqs.push({ subject_id: onereq.replace(/'/g, ''), title: '', fulfilled: false });
             } else {
+              let subRequirement = getClassInfo(onereq);
+              if (_this.firstAppearance >= -1) {
+                const allSemesterSubjects = _this.$store.state.roads[_this.$store.state.activeRoad].contents.selectedSubjects;
+                let allPreviousSubjects = _this.flatten(allSemesterSubjects.slice(0, _this.firstAppearance));
+                subRequirement.fulfilled = _this.reqFulfilled(onereq, allPreviousSubjects);
+              } else {
+                subRequirement.fulfilled = true;
+              }
               parsedReq.reqs.push(getClassInfo(onereq));
             }
           } else {
             parsedReq.reqs.push(parseReqs(onereq));
           }
         }
+
         if (connectionType === '/') {
           parsedReq.connectionType = 'any';
+          parsedReq.fulfilled = parsedReq.reqs.some((req) => req.fulfilled);
         } else if (connectionType === ',') {
           parsedReq.connectionType = 'all';
+          parsedReq.fulfilled = parsedReq.reqs.every((req) => req.fulfilled);
         }
+
         function sortOrder (req) {
           if (req.reqs !== undefined) {
             return 0;
@@ -370,6 +400,7 @@ export default {
             return req.subject_id + ' ' + req.title;
           }
         }
+
         if (parsedReq.reqs.length === 2) {
           if (parsedReq.connectionType === 'any') {
             parsedReq.subject_id = getReqTitle(parsedReq.reqs[0]);
@@ -390,7 +421,9 @@ export default {
             parsedReq.expansionDesc = 'Select all:';
           }
         }
+
         const connectionMatch = /(and|or)/.exec(parsedReq.subject_id);
+
         if (connectionMatch !== null) {
           const connectionIndex = connectionMatch.index;
           const firstPart = parsedReq.subject_id.substring(0, connectionIndex).replace(/\s/g, '');
@@ -398,9 +431,12 @@ export default {
           parsedReq.subject_id = firstPart;
           parsedReq.title = secondPart + ' ' + parsedReq.title;
         }
+
         return parsedReq;
       }
+
       const rList = parseReqs(requirements);
+      console.log(rList);
       rList.topLevel = true;
       return rList;
     },
